Enabling bitcode for release builds
We're no longer using the old `packForXcode` task which we had passed a configuration to that determined if bitcode was embedded or just marked. As of Kotlin 1.3.20, embedding bitcode is done by default for iOS framework targets. The `:transport:assembleMessengerTransportReleaseXCFramework` task from the XCFramework plugin appears to be building them correctly.