GMDX-383 partial
Refactor the token store. The refactoring of the MobileMessenger will include the ability to pass in a TokenStore. Note that there will be additional changes that make this change make more sense... we don't instantiate the DefaultTokenStore in the MobileMessenger classes as it's done right now moving forward, that's just to get this refactor in a PRable state.