Rename and refactor
This PR renames MobileMessenger to MessengerTransport. It also modifies the semantics of interacting with the MessengerTransport to allow creation and interaction from the common codebase rather than just from native. The deployment configuration is now passed into the constructor of the MessengerTransport, simplifying the calling of the other methods on the object.