Mitigating new socket behavior seen in iOS 16.4
[MTSDK-183]
In iOS 16.4, when the network becomes unavailable, such as when the device is put into airplane mode, the NSURLSessionWebSocketTask now immediately completes with an error. This happens without the socket closing first. Furthermore, the completion handler for sendPingWithPongReceiveHandler happens on something other than the main thread. To mitigate, we can listen for the socket task completion as well as ensure that our external facing listeners from the SDK are dispatched on Main.